requirements.txt
google-genai>=1.10.0
pydantic>=2.7.0
python-dotenv>=1.0.1
pandas>=2.2.2
